Move LocalAddressVisitor to a separate file
